[FIX] 修复批量删除,重复删除子任务报错

+++ src/main/java/io/choerodon/agile/app/service/impl/IssueOperateServiceImpl.java
@@ -5,6 +5,7 @@ import io.choerodon.agile.api.vo.BatchUpdateFieldStatusVO;
 import io.choerodon.agile.app.service.IssueOperateService;
 import io.choerodon.agile.app.service.IssueService;
 import io.choerodon.agile.app.service.UserService;
+import io.choerodon.agile.infra.mapper.IssueMapper;
 import io.choerodon.core.exception.CommonException;
 import io.choerodon.core.oauth.DetailsHelper;
 import org.hzero.boot.message.MessageClient;
@@ -14,6 +15,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -31,6 +33,8 @@ public class IssueOperateServiceImpl implements IssueOperateService {
     private MessageClient messageClient;
     @Autowired
     private UserService userService;
+    @Autowired
+    private IssueMapper issueMapper;
 
     @Async
     @Override
@@ -51,10 +55,16 @@ public class IssueOperateServiceImpl implements IssueOperateService {
                 batchUpdateFieldStatusVO.setUserId(userId);
                 batchUpdateFieldStatusVO.setProcess(progress);
                 messageClient.sendByUserId(userId, messageCode, JSON.toJSONString(batchUpdateFieldStatusVO));
+                // 查询子任务ids
+                List<Long> subList = new ArrayList<>();
+                subList.addAll(issueMapper.selectSubListByIssueIds(projectId,issueIds));
                 int i = 0;
                 for (Long issueId : issueIds) {
                     i++;
-                    issueService.deleteIssue(projectId, issueId);
+                    // 删除任务会附带删除子任务,因此需要判断问题是否被删除ss
+                    if (!subList.contains(issueId)) {
+                        issueService.deleteIssue(projectId, issueId);
+                    }
                     if (i % incremental == 0) {
                         batchUpdateFieldStatusVO.setProcess((i * 1.0) / issueIds.size());
                     }
